+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # 8.5.0
+## ADDED
+- Custom icons for the hold action attribute can now be defined in the Editor preferences -> Misc.
+- Icon picker to the hold action attribute  #435
+- ENH_fnc_iconPicker, an simple Icon Picker
+
 ## CHANGED
 - Fixed delete crew tool would not work on UAVs/UAV Turrets (Thanks hypoxia125) #433
 - Fixed that unit traits and attributes would conflict with ACE (Thanks Nomas-X for the detailed issue) #436
-- Increased font size of hold action icon combos
-- Hold action icon combos now only show icon name and not the entire path #435
 
 # 8.4.0
 ## ADDED